compiler warning fixes and cosmetics

+++ src/modules/au/test_au_dbupdate.c
@@ -191,7 +191,7 @@ main (int argc, char *argv[])
   whois = LU_whois_init(whois_server, whois_port, 30);
 
   /* initialise AU */
-  AU_init(ctx, (int)whois);
+  AU_init(ctx, (long)whois);
 
   /* read object */
   obj = read_object(input_file);
